requirements.txt
annotated-types==0.7.0
anyio==4.7.0
beautifulsoup4==4.12.3
click==8.1.7
fastapi==0.115.6
google==3.0.0
grpcio==1.68.1
grpcio-tools==1.68.1
h11==0.14.0
idna==3.10
packaging==24.2
pipdeptree==2.24.0
prettytable==3.12.0
protobuf==5.29.1
pydantic==2.10.3
pydantic_core==2.27.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.2
schedule==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
treelib==1.7.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==4.2
ultipa==4.5.2
uvicorn==0.32.1
wcwidth==0.2.13
